+++ CHANGELOG.md
@@ -2694,3 +2694,4 @@
 - [2026-07-11 12:46] fix(core): fix memory leak by cancelling core subscription
 - [2026-07-11 12:50] style(cache): harmonize light and dark theme palette for cache
 - [2026-07-11 13:22] chore(i18n): update dependencies and cleanup unused imports in i18n
+- [2026-07-11 13:54] test(network): implement widget integration tests for network